MathCam/backend/ml_models/ocr/image_processor.py
import io
import cv2
import numpy as np
from PIL import Image
import tensorflow as tf

# Ce module sera chargé de traiter les images d'entrée et d'extraire les équations via OCR

# Fonction fictive qui sera remplacée par un vrai modèle ML
async def extract_equation(image_file):
    """
    Extrait une équation mathématique à partir d'une image en utilisant un modèle OCR.
    
    Args:
        image_file: Le fichier image téléchargé contenant l'équation mathématique
    
    Returns:
        str: L'équation mathématique extraite sous forme de texte
    """
    try:
        # Lire l'image en mémoire
        contents = await image_file.read()
        nparr = np.frombuffer(contents, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        
        # Prétraitement de l'image
        preprocessed_img = preprocess_image(img)
        
        # Extraction de l'équation (à implémenter avec un vrai modèle ML)
        equation = recognize_math_symbols(preprocessed_img)
        
        return equation
        
    except Exception as e:
        raise Exception(f"Erreur lors du traitement de l'image: {str(e)}")

def preprocess_image(img):
    """
    Prétraite l'image pour améliorer la reconnaissance de caractères
    """
    # Conversion en niveaux de gris
    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
    
    # Appliquer un flou gaussien pour réduire le bruit
    blurred = cv2.GaussianBlur(gray, (5, 5), 0)
    
    # Binarisation adaptative
    thresh = cv2.adaptiveThreshold(blurred, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, 
                                  cv2.THRESH_BINARY_INV, 11, 2)
    
    # Dilatation pour connecter les composants proches
    kernel = np.ones((2, 2), np.uint8)
    dilated = cv2.dilate(thresh, kernel, iterations=1)
    
    return dilated

def recognize_math_symbols(img):
    """
    Reconnaît les symboles mathématiques dans l'image prétraitée
    Note: Cette fonction est un placeholder pour le vrai modèle ML
    """
    # À remplacer par un vrai modèle de reconnaissance de caractères mathématiques
    # Pour l'instant, retournons une équation d'exemple
    return "2x + 3 = 7"

# Fonction pour charger un modèle ML préentraîné
def load_model():
    """
    Charge le modèle ML préentraîné pour la reconnaissance de symboles mathématiques
    """
    # À implémenter avec TensorFlow ou un autre framework ML
    pass 