MathCam/backend/requirements.txt
sympy==1.11.1
numpy==1.24.3
pillow==9.5.0
tensorflow==2.12.0
keras==2.12.0
opencv-python==4.7.0.72
scikit-image==0.20.0
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7 